Omit unnecessary code from bundled SQLite build
Following the explanations from SQLite's [recommended compile-time options](https://www3.sqlite.org/compile.html),
this adds a Cargo feature to libsqlite-sys so that the bundled SQLite can be better tailored to the specific needs
of rusqlite. The feature is enabled by default and rusqlite does explicitly not request it, so that if any other
crate pulls libsqlite-sys into the dependency, no features are removed.

As a first step, this commit
* adds `SQLITE_OMIT_PROGRESS_CALLBACK` as rusqlite does not use `sqlite3_progress_handler`
* and sets `SQLITE_DEFAULT_MEMSTATUS` to `0` as rusqlite does not use `sqlite3_status`.