Improve API VTab::best_index: Separate mutable and immutable data
Mixing mutable and immutable data has several problems, one is that
you cannot loop over `IndexConstraintIter` while setting respective
`IndexConstraintUsage`.

This makes it hard to implement `VTab::best_index`.

Another problem is that it is confusing, as programmer do not know which
part is data for reading and which part is required to be set for the
`VTab::best_index` implementation to works.

This commit separates mutable from the immutable, by passing additional
param `constraint_usages` and makes `info` to be immutable, and requires
the implementation to returns a `BestIndex` so that these fields will
always be explicitly initialized.

param `constraint_usages` also automatically sets each
`IndexConstraintUsage` to some sane values before passing it to
`VTab::best_index` implementation.

Signed-off-by: Jiahao XU <Jiahao_XU@outlook.com>