Fix build for android x86_64
Adds a temporary workaround for [an issue] with the Rust compiler and Android when
compiling for x86_64 devices.

The Android NDK used to include `libgcc` for unwind support (which is required by Rust
among others). From NDK r23, `libgcc` is removed, replaced by LLVM's `libunwind`.
However, `libgcc` was ambiently providing other compiler builtins, one of which we
require: `__extenddftf2` for software floating-point emulation. This is used by SQLite
(via the `rusqlite` crate), which defines a `LONGDOUBLE_TYPE` type as `long double`.

Rust uses a `compiler-builtins` crate that does not provide `__extenddftf2` because
[it involves floating-point types that are not supported by Rust][unsupported]. For
some reason, they _do_ export this symbol for `aarch64-linux-android`, but they do not
for `x86_64-linux-android`. Thus we run into a problem when trying to compile and run
the SDK on an x86_64 emulator.

The workaround comes from [this Mozilla PR]: we tell Cargo to statically link the
builtins from the Clang runtime provided inside the NDK, to provide this symbol.

[an issue]: https://github.com/rust-lang/rust/issues/109717
[this Mozilla PR]:https://github.com/mozilla/application-services/pull/5442
[unsupported]: https://github.com/rust-lang/compiler-builtins#unimplemented-functions

This fix was copied from: https://github.com/Electric-Coin-Company/zcash-android-wallet-sdk/commit/1bf2f848373bee833a11fbb8cd9b570218ac71bb

Fixes #1380

Co-authored-by: Jack Grigg <jack@electriccoin.co>