Replace use of function pointers with `Fn*` traits
This commit replaces all instances of methods accepting `fn(_) -> _`
with `F: Fn(_) -> _`, for greater flexibility.

A new internal abstraction is introduced to facilitate this change:
`ThinBoxAny`, a type similar to `Option<Box<dyn Any>>` but stored in a
single pointer in order to avoid inflating the size of the `Connection`
type too much.

There are two exceptions, the deprecated `trace` and `profile`
functions. Users who need this flexibility should migrate to `trace_v2`
instead.

This change is, unfortunately, breaking. There are two ways in which it
breaks code:
+ Code that uses something like `.wal_hook(None)` will now get errors
  because Rust cannot infer the type parameter of `F`. Users must now
  write `.wal_hook(None::<fn(&Wal, _) -> _>)`.
+ Code that uses something like `.wal_hook(|hook, _| {})` will now fail,
  because Rust’s type inference of closures in the presence of HRTBs is
  limited. Instead, users should write `.wal_hook(|hook: &Wal, _| {})`.

The change is conservative for now, generally requiring `Fn` and `Sync`
even when this may not be strictly necessary. It is likely many of these
bounds can be relaxed in future.

Fixes: #977